requirements.txt
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
colorama>=0.4.4
tabulate>=0.8.9